add question to answer range my age

+++ db/migrate/20160920083452_create_stages.rb
@@ -1,8 +1,8 @@
 class CreateStages < ActiveRecord::Migration
   def change
     create_table :stages do |t|
-      t.integer :from_age
-      t.integer :to_age
+      t.float :from_age
+      t.float :to_age
 
       t.timestamps null: false
     end
